Switched from callbacks to using GLFW functions for getting mouse pos to fix imgui not properly receiving mouse input. Scroll doesn't have function only callback so need to investigate if that's an issue with imgui or not. Can now set how many chunks are rendering using a slider in imgui.

+++ src/voxel_world.hpp
@@ -27,7 +27,7 @@ private:
 public:
   VoxelWorld(int chunk_size, int size_x, int size_y, int size_z);
 
-  size_t chunks_to_render;
+  int chunks_to_render;
 
   void generate(unsigned int seed);
   void render(const Shader &shader);
